Closer to completing the content item object
These late nigths are going to kill me (yes 1 am is late, especially when
you have kills that are up at 6am sharp). Theres not really enough time to
get a lot of work done in one shot. Anyway, I digress; Implemented a little
more of the content items object and some of the is classes that hang off
of it.