Minor changes to the over all design of the API
I gave some thought to how the end developer will access the API, how they
will navigate between objects, access data, access the cache and what they
will be able to do and how. I've not finished all of the changes that I
need to make for this but its a step in the right direction.