A little further with UI first
My ultimate goal is to have the admin section built with the same exact API
that end developers will use. The UI in the admin section should not make
use of any functionality that is not encapsulated in the API. To that end
before I get to deep into fleshing out the API I'd like to see how it needs
to be used. This is the reason for building the UI at the same time as the
API.

This project is getting so huge, even most of the things in this project
are just bare bones mock up I still haven't mocked up 90 percent of what
will eventually be in here. Please help anyone!

Signed-off-by: Anthony Romero <dn4378@yahoo.com>